server/requirements-dev.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-environ
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
requests
